refactor: move ext files to utils
