+++ ui/src/main/kotlin/jp/co/yumemi/android/code_check/ui/features/details/repo/RepoDetailsScreen.kt
@@ -29,7 +29,6 @@ import jp.co.yumemi.android.code_check.ui.components.error.title
 import jp.co.yumemi.android.code_check.ui.components.repo.RepoDetailsElementItem
 import jp.co.yumemi.android.code_check.ui.components.repo.RepoDetailsHeader
 import jp.co.yumemi.android.code_check.ui.core.Contract
-import jp.co.yumemi.android.code_check.ui.core.elevation
 import jp.co.yumemi.android.code_check.ui.core.handle
 import jp.co.yumemi.android.code_check.ui.core.render
 import jp.co.yumemi.android.code_check.ui.primitives.Gray
