+++ ui/src/main/kotlin/jp/co/yumemi/android/code_check/ui/features/search/results/SearchRepoResultsFragment.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
-import jp.co.yumemi.android.code_check.ui.core.compose
+import jp.co.yumemi.android.code_check.ui.utils.compose
 import jp.co.yumemi.android.code_check.ui.core.contract
 
 @AndroidEntryPoint
